refactor site thumbnail to move mshots logic into it's own file
